feat(make): Add SRCDIR to Makefile
The current Makefile expects the source files (`index.md` and
`template.html`) in the `demo` directory. If another source directory is
desired, the Makefile must be modified. This is can be the case if
multiple `index.md` files are being maintained.

This change adds the `SRCDIR` variable and sets the default to `demo`.
This allows for default behavior. If another source directory should be
used the following command line works:

    SRCDIR=my-web-site make

This will cause the Makefile to look for `index.md` and `template.html`
in the `my-web-site` directory.